refactor: src/test/java/seedu/address/logic/commands/
